requirements.txt
# ============================================================================
# GENDER VOICE DETECTION - MLOps Project Dependencies
# ============================================================================
# Python 3.9+ required
# Last updated: December 2025
# ============================================================================

# Deep Learning & ML
# ----------------------------------------------------------------------------
tensorflow==2.16.1                # Core ML framework untuk RNN/LSTM/GRU (upgraded for typing-extensions compatibility)
scikit-learn==1.3.2               # Machine learning utilities & metrics
numpy==1.26.4                     # Numerical computing (compatible with TF 2.16)
pandas==2.1.3                     # Data manipulation

# Audio Processing
# ----------------------------------------------------------------------------
librosa==0.10.1                   # Audio analysis & feature extraction
soundfile==0.12.1                 # Audio file I/O
noisereduce==3.0.0                # Noise reduction preprocessing
pydub==0.25.1                     # Audio manipulation & conversion

# Web Framework & API
# ----------------------------------------------------------------------------
fastapi==0.104.1                  # Modern web framework untuk REST API
uvicorn[standard]==0.24.0         # ASGI server
python-multipart==0.0.6           # Form data parsing untuk file upload
pydantic==2.5.0                   # Data validation
pydantic-settings==2.1.0          # Settings management

# UI & Visualization
# ----------------------------------------------------------------------------
gradio==4.7.1                     # Interactive ML web interface
matplotlib==3.8.2                 # Plotting library
seaborn==0.13.0                   # Statistical data visualization
plotly==5.18.0                    # Interactive plots

# MLOps & Experiment Tracking
# ----------------------------------------------------------------------------
mlflow==2.8.1                     # Experiment tracking & model registry
dagshub==0.3.1                    # MLflow hosting & collaboration
dvc==3.30.1                       # Data version control
dvc-gdrive==2.20.0                # DVC remote storage untuk Google Drive

# Utilities
# ----------------------------------------------------------------------------
python-dotenv==1.0.0              # Environment variables management
pyyaml==6.0.1                     # YAML configuration files
tqdm==4.66.1                      # Progress bars
loguru==0.7.2                     # Advanced logging
python-jose[cryptography]==3.3.0  # JWT untuk admin authentication
passlib[bcrypt]==1.7.4            # Password hashing
reportlab==4.0.7                  # PDF generation untuk reports

# Testing & Quality
# ----------------------------------------------------------------------------
pytest==7.4.3                     # Testing framework
pytest-cov==4.1.0                 # Code coverage
pytest-asyncio==0.21.1            # Async testing support
black==23.11.0                    # Code formatter
flake8==6.1.0                     # Linting
mypy==1.7.1                       # Type checking

# Async & Background Tasks
# ----------------------------------------------------------------------------
celery==5.3.4                     # Distributed task queue untuk async training
redis==5.0.1                      # Message broker untuk Celery (optional)

# Additional Tools
# ----------------------------------------------------------------------------
requests==2.31.0                  # HTTP library
aiofiles==23.2.1                  # Async file operations
pillow==10.1.0                    # Image processing (untuk visualizations)
